# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Tarefas e Usuários",
    "version": "1.0.0",
    "description": "API para registro de usuários, login, cadastro e consulta de tarefas."
  },
  "paths": {
    "/api/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "400": { "description": "Usuário já existe ou dados inválidos" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido, retorna token" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
  "/api/tasks": {
      "post": {
        "summary": "Cadastrar tarefa",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "priority": { "type": "string", "enum": ["baixa", "media", "alta"] },
                  "dueDate": { "type": "string", "format": "date-time" }
                },
                "required": ["title", "priority", "dueDate"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tarefa criada" },
          "400": { "description": "Erro de validação" },
          "401": { "description": "Token não informado ou inválido" }
        }
      },
      "get": {
        "summary": "Consultar tarefas do usuário autenticado",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de tarefas" },
          "401": { "description": "Token não informado ou inválido" }
        }
      }
    },
    "/api/tasks/{id}": {
      "put": {
        "summary": "Atualizar tarefa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "priority": { "type": "string", "enum": ["baixa", "media", "alta"] },
                  "dueDate": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tarefa atualizada com sucesso" },
          "404": { "description": "Tarefa não encontrada" },
          "401": { "description": "Token não informado ou inválido" }
        }
      },
      "delete": {
        "summary": "Remover tarefa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Tarefa removida com sucesso" },
          "404": { "description": "Tarefa não encontrada" },
          "401": { "description": "Token não informado ou inválido" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}